mbed version working with a new board revision.
I wanted to get something up and running in my apartment without
implementing all of the BLE stuff in v2, so I backported the inverted
UART TX pin code to the SDK which mbed uses.

The larger Fanstel modules (BT832) have *much* better Bluetooth range,
and they're about a buck cheaper, so I'm using them going forward. The
tiny BC832 is *super* cool but otherwise has no advantages in this use
case, so I'll hang onto them for other projects. The BT832 has built-in
inductors for the DCDC converter, as do the other boards I have, so I
switched it on for better power efficiency.

The rest was a big struggle. I kept getting strange behavior with the
newer boards which use MOSFETs for switching instead of optocouplers.
The A/C system would get stuck in an error condition, or only address
one packet to the secondary thermostat and then never send any more even
if it responded, unless I kept 0x20 set in the first field of my
broadcasts (which seems to indicate "re-poll the bus for thermostats").
After a *long* day or so with an oscilloscope, I discovered that pulling
the bus hard to ground is ignored by the thermostat(s) but not the
ceiling unit. I always saw that the thermostat and ceiling unit pulled
the bus *almost* to ground but not quite, but I ignored it because
things seemed to work. But, looking at pad T17 on a thermostat, I saw
pulses from the ceiling unit and the other thermostat reflected, but
nothing from my board.

There's a 220Ω resistor which seems to be on the thermostat's transmit
path, and I think it's part of some kind of short circuit handling: if
the bus goes all the way to ground, the thermostats act like it's idle.
I threw a resistor between the bus and the MOSFET, and suddenly the
thermostat hears us. Sweet. The 0x20 worked because it got reflected by
the ceiling unit on the next cycle, so the primary thermostat would
always address the secondary even if it never heard back. I'm super
happy to have fixed it electrically instead of abusing this bit.