Fix BLE on the BC832/BT832.
I take back what I said in the last commit: the Bluetooth range on the
BC832 is fine, I just did a poor job pinpointing the problem. I was able
to connect to both boards from Apple devices, but not from a Raspberry
Pi. Nothing to do with distance!

This change lets both Apple and Linux devices discover the board.